+++ upi-qr-code-payment-for-woocommerce.php
@@ -3,14 +3,14 @@
  * Plugin Name: UPI QR Code Payment for WooCommerce
  * Plugin URI: https://wordpress.org/plugins/upi-qr-code-payment-for-woocommerce/
  * Description: It enables a Woocommerce site to accept payments through UPI apps like BHIM, Google Pay, Paytm, PhonePe or any Banking UPI app. Avoid payment gateway charges.
- * Version: 1.1.2
+ * Version: 1.1.6
  * Author: Sayan Datta
  * Author URI: https://www.sayandatta.in
  * License: GPLv3
  * Text Domain: upi-qr-code-payment-for-woocommerce
  * Domain Path: /languages
  * WC requires at least: 3.1
- * WC tested up to: 4.2
+ * WC tested up to: 4.5
  * 
  * UPI QR Code Payment for WooCommerce is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -39,7 +39,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 $consts = array(
-    'UPI_WOO_PLUGIN_VERSION'       => '1.1.2', // plugin version
+    'UPI_WOO_PLUGIN_VERSION'       => '1.1.6', // plugin version
     'UPI_WOO_PLUGIN_BASENAME'      => plugin_basename( __FILE__ ),
 	'UPI_WOO_PLUGIN_DIR'           => plugin_dir_url( __FILE__ ),
 	//'UPI_WOO_PLUGIN_ENABLE_DEBUG'  => true